+++ R/occCitePrint.R
@@ -1,6 +1,7 @@
 #' @title Print occCite citation object
 #'
-#' @description Prints formatted citations for occurrences
+#' @description Prints formatted citations for occurrences and main
+#' packages used (i.e. base, occCite, rgbif, and/or BIEN).
 #'
 #' @param x An object of class \code{\link{occCiteCitation}}
 #'
@@ -8,6 +9,9 @@
 #' citation document is produced
 #'
 #' @return A text string with formatted citations
+#'
+#' @import bibtex
+#' @import bib2df
 #'
 #' @examples
 #'
@@ -31,24 +35,106 @@ print.occCiteCitation <- function(x, ...) {
 
   stopifnot(inherits(x, "occCiteCitation"))
 
+  # Function to generate package citations
+  packageCitations <- function(packagesUsed){
+    bibtex::write.bib(packagesUsed, "temp.bib", verbose = F)
+    temp <- bib2df::bib2df("temp.bib", separate_names = T)
+
+    packageCitations <- vector(mode = "character", length = nrow(temp))
+    for(j in 1:nrow(temp)){
+      i <- temp[j,]
+      if("R Core Team" %in% i$AUTHOR[[1]]["full_name"]){
+        packageCitations[j] <- paste0(i$AUTHOR[[1]]['full_name'], ". ", "(",i$YEAR, "). ",
+                                      i$TITLE, ". ", i$ORGANIZATION, ", ",
+                                      i$ADDRESS, ". ", i$URL, ".")
+      } else if(i$CATEGORY=="ARTICLE"){
+        packageCitations[j] <- paste0(paste(apply(i$AUTHOR[[1]],
+                                                  MARGIN = 1,
+                                                  FUN = function(x) paste0(x[4], ", ",
+                                                                           substring(x[2], 1, 1), ".",
+                                                                           if(!is.na(x[3])){
+                                                                             paste0(substring(x[3], 1, 1), ".")})),
+                                            collapse = ", "),
+                                      " (",i$YEAR, "). ",
+                                      i$TITLE, ". ",
+                                      i$JOURNAL, if(!any(is.na(i$VOLUME), is.na(i$PAGES))){paste0(" ")},
+                                      if(!is.na(i$VOLUME)){paste0(i$VOLUME, ": ")},
+                                      if(!is.na(i$PAGES)){paste0(i$PAGES)},
+                                      "."
+        )
+      } else if(grepl("R package", i$NOTE)){
+        packageCitations[j] <- paste0(paste(apply(i$AUTHOR[[1]],
+                                                  MARGIN = 1,
+                                                  FUN = function(x) paste0(x[4], ", ",
+                                                                           substring(x[2], 1, 1), ".",
+                                                                           if(!is.na(x[3])){
+                                                                             paste0(substring(x[3], 1, 1), ".")})),
+                                            collapse = ", "),
+                                      " (",i$YEAR, "). ",
+                                      i$TITLE, ". ",
+                                      i$NOTE, ". ",
+                                      i$URL, ".")
+      }
+      else{
+        packageCitations[j] <- paste0(paste(apply(i$AUTHOR[[1]],
+                                                  MARGIN = 1,
+                                                  FUN = function(x) paste0(x[4], ", ",
+                                                                           substring(x[2], 1, 1), ".",
+                                                                           if(!is.na(x[3])){
+                                                                             paste0(substring(x[3], 1, 1), ".")})),
+                                            collapse = ", "),
+                                      " (",i$YEAR, "). ",
+                                      i$TITLE, ". ",
+                                      i$ORGANIZATION, ", ",
+                                      i$ADDRESS, ". ",
+                                      i$URL, ".")
+      }
+    }
+
+    file.remove("temp.bib")
+    packageCitations <- sort(packageCitations)
+    return(packageCitations)
+  }
+
+  # Combine citations
   if(bySpecies){
     for (i in 1:length(x$occCitationResults)){
+      # Get list of packages used and get citations
+      packCit <- c("base", "occCite")
+      packages <- unique(x$occCitationResults[[i]]$occSearch)
+      if("GBIF" %in% packages) packCit <- c(packCit, "rgbif")
+      if("BIEN" %in% packages) packCit <- c(packCit, "BIEN")
+      packCit <- packageCitations(packCit)
+
+      # Print results
       cat(paste("Species:", names(x$occCitationResults)[[i]], "\n\n"))
       singleSpRecord <- x$occCitationResults[[i]]
       singleSpRecord <- singleSpRecord[order(singleSpRecord$Citation),]
-      cat(paste0(singleSpRecord$Citation,
-                " Accessed via ", singleSpRecord$occSearch,
-                " on ", singleSpRecord$`Accession Date`, "."),
-          sep = "\n")
+      recordCitations <- paste0(singleSpRecord$Citation,
+             " Accessed via ", singleSpRecord$occSearch,
+             " on ", singleSpRecord$`Accession Date`, ".")
+      allCitations <- sort(c(packCit, recordCitations))
+      cat(allCitations, sep = "\n")
       cat(paste("\n"))
     }
   } else {
-    allCitations <- do.call(rbind, x$occCitationResults)
-    allCitations <- allCitations[order(allCitations$Citation),]
-    allCitations <- unique(allCitations[,c("occSearch", "Citation", "Accession Date")])
-    cat(paste0(allCitations$Citation,
-              " Accessed via ", allCitations$occSearch,
-              " on ", allCitations$`Accession Date`, "."),
-        sep = "\n")
+    recordCitations <- do.call(rbind, x$occCitationResults)
+
+    # Get list of packages used and get citations
+    packCit <- c("base", "occCite")
+    packages <- unique(recordCitations$occSearch)
+    if("GBIF" %in% packages) packCit <- c(packCit, "rgbif")
+    if("BIEN" %in% packages) packCit <- c(packCit, "BIEN")
+    packCit <- packageCitations(packCit)
+
+    # Print results
+    recordCitations <- recordCitations[order(recordCitations$Citation),]
+    recordCitations <- unique(recordCitations[,c("occSearch", "Citation", "Accession Date")])
+    recordCitations <- paste0(recordCitations$Citation,
+              " Accessed via ", recordCitations$occSearch,
+              " on ", recordCitations$`Accession Date`, ".")
+    allCitations <- sort(c(packCit, recordCitations))
+    cat(allCitations, sep = "\n")
+    cat(paste("\n"))
   }
 }