+++ NEWS.md
@@ -6,7 +6,7 @@
 # occCite 0.2.0
 
 * Updated to reflect new object structures returned by searches with update `rgbif`, version 3.1.
-* Examples update to permit testing when possible.
+* Examples updated to permit testing when possible.
 
 # occCite 0.1.0